Added CombinableItem & Combination(s)
This is a more significant change in the way that we do combinations.
The standard is now to store them in a

```
Map<Set<CombinableItem>, Item>
```

where `CombinableItem` extends `Item`. The hash (and equality) of a
`Set` is order-agnostic, so it works well as a key for multiple items.
This limits us in being able to have multiple copies of one item used in
a combination, but this isn't designed to make Minecraft, so it should be
fine for now.