Added Situation-related skeleton
Still rather useless until we add some kind of execution. Also, adding multiple `SituationNode`s with children of their own is problematic with the current builder format. I'll think of some way to improve that later, and am open to suggestions. Likely just an addMany() with the ability to easily instantiate SituationNodes