Cleanup, refactoring, minor test additions
- Hashed out how `Action`s work a little more, and put them into use
- Removed rogue `prompt` argument in `respondToInteraction` that didn't
make sense
- Other minor cleanup and refactoring
- Added `Confirmation` as a generic confirmation dialogue. Not sure if I
want this to be a separate class, as I have made it, or if it makes
sense to model it as part of something that looks like `ConsoleActors`
(which is now named `ConsolePrompt`). I'm thinking a static factory
method. This is assuming that we have a few more things like this, which
sit at a slightly higher level than those in `ConsolePrompt` in that
they act as `Interactive` objects, but their entire purpose is to
simplify basic, repetitive tasks like popping confirmation modals that
should have a resulting action.